+++ src/main/java/ak/dev/khi_backend/user/configs/SecurityConfig.java
@@ -110,6 +110,7 @@ public class SecurityConfig {
                                 "/api/v1/about/partners/**",
                                 "/api/v1/settings/social/**",
                                 "/api/v1/donations/type-cards/**",
+                                "/api/v1/book-genres/**",
                                 "/api/v1/nav-menu/**"
                         ).hasAnyRole("ADMIN", "SUPER_ADMIN")
                         .requestMatchers(HttpMethod.PUT,
@@ -119,6 +120,7 @@ public class SecurityConfig {
                                 "/api/v1/settings/social/**",
                                 "/api/v1/donations/settings",
                                 "/api/v1/donations/type-cards/**",
+                                "/api/v1/book-genres/**",
                                 "/api/v1/site-settings",
                                 "/api/v1/nav-menu/**"
                         ).hasAnyRole("ADMIN", "SUPER_ADMIN")
@@ -128,6 +130,7 @@ public class SecurityConfig {
                                 "/api/v1/about/partners/**",
                                 "/api/v1/settings/social/**",
                                 "/api/v1/donations/type-cards/**",
+                                "/api/v1/book-genres/**",
                                 "/api/v1/nav-menu/**"
                         ).hasAnyRole("ADMIN", "SUPER_ADMIN")